problems/youtube/requirements.txt
httpx>=0.27.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
pandas >= 2.2.3
tqdm >= 4.66.5
